Do schema migration more sensibly
 * Keep v0 schema in .sql scripts
 * Run schema scripts if version == 0
 * Do schema upgrades if v < current, so new db will get v0 schema then upgrade
